requirements.txt
pytesseract
pyautogui
psutil
pynput
opencv-python
pillow
pydirectinput
tensorflow
numpy
opencv-python
